rename `_device.py` to `_legacy_device.py`
