+++ pennylane/devices/_legacy_device.py
